Fixed compilation error
Former-commit-id: 14644cf9c03e8b0fc472d683d57a92dd5b78f223

+++ hq-server/src/main/java/org/hyperic/hq/appdef/server/session/AgentManagerImpl.java
@@ -163,7 +163,7 @@ public class AgentManagerImpl implements AgentManager, ApplicationContextAware {
      */
     @Transactional(readOnly = true)
     public int getAgentCountUsed() {
-        return agentDao.countUsed();
+        return (int)agentDao.countUsed();
     }
     
     /**
